# Reduce rb_type() calling in `Oj.dump`