This patch will use switch statement instead to reduce times of rb_type() calling with if statement.

−               | before  | after    | result
--               | --      | --       | --
Oj.dump          | 1.175M  | 1.191M   | 1.014x

### Environment
- MacBook Air (M1, 2020)
- macOS 12.0
- Apple M1
- Ruby 3.0.2

### Before
```
Warming up --------------------------------------
             Oj.dump   119.325k i/100ms
Calculating -------------------------------------
             Oj.dump      1.175M (± 0.3%) i/s -      5.966M in   5.079028
```

### After
```
Warming up --------------------------------------
             Oj.dump   117.261k i/100ms
Calculating -------------------------------------
             Oj.dump      1.191M (± 0.5%) i/s -      5.980M in   5.021336s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json =<<-EOF
{
  "$id": "https://example.com/person.schema.json",
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "title": "Person",
  "type": "object",
  "properties": {
    "firstName": {
      "type": "string",
      "description": "The person's first name."
    },
    "lastName": {
      "type": "string",
      "description": "The person's last name."
    },
    "age": {
      "description": "Age in years which must be equal to or greater than zero.",
      "type": "integer",
      "minimum": 0
    }
  }
}
EOF

data = Oj.load(json, symbol_keys: true)
Benchmark.ips do |x|
  x.report("Oj.dump") { Oj.dump(data) }
end
```